requirements.txt
Django
mozilla-django-oidc
python-decouple